docker/ocr-minimal/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pillow>=10.0.0
pydantic>=2.0.0
httpx>=0.25.0
